+++ osa6/redux-anecdotes/src/components/AnecdoteList.jsx
@@ -1,0 +1,30 @@
+import { useSelector, useDispatch } from 'react-redux'
+import { addVote } from '../reducers/anecdoteReducer'
+
+
+const AnecdoteList = () => {
+  const anecdotes = useSelector(state => state)
+  const dispatch = useDispatch()
+
+  const vote = (id) => {
+    dispatch(addVote(id))
+  }
+
+  return (
+    <div style={{marginTop: "50px"}}>
+      {anecdotes.map(anecdote =>
+        <div key={anecdote.id}>
+          <div>
+            {anecdote.content}
+          </div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => vote(anecdote.id)}>vote</button>
+          </div>
+        </div>
+      )}
+    </div>
+  )
+}
+
+export default AnecdoteList
